User control over permutation type
Added support for permutations in which values are repeated but all
input values must be present.  Included "allow reps" and "require all"
in the input grammar to enable a problem-specification file to
indicate the type of permutation to use.

Made ParsedInput records implement the Show interface.  For
convenient, the output is written in a form that's valid for input.