Improved performance with new tree code
At Kei Davis's suggestion, rewrote the tree-generation code to first
generate Schroeder paths, then map each Schroder path to a tree.  The
intention was to save memory by not having to retain all tree
structures for the duration of the program's execution, but it turns
out that this also makes the program run significantly faster.