Corrected the safe-division operator
The safeDiv function already checked for division by zero but still
threw an exception when dividing minBound by -1.  This version handles
that special case, too.  In addition, when converting to a string,
division is now represented as "floor(a / b)" instead of just "a / b"
to make it clear that we're discussing integer division.