Riffle-shuffle each list of stack operators
The goal of this change is to improve the program's average
performance.  The idea is that multiple expression structures will
satisfy most inputs, but a few inputs require unique structures (e.g.,
a op (b op (c op d))).  Bringing these special cases nearer to the
front of the list should improve performance while not adversely
affecting the "easy" cases.