Boolean system if inputs *and* outputs are Boolean
Previously, only inputs were checked for being exclusively 0 and 1.
If so, findexpr defaults to using NOT as the only unary operator and
AND, OR, XOR, NAND, and NOR as the binary operators.  In this
revision, both inputs *and* outputs are considered.  Otherwise,
findexpr will search forever to find a map from {0, 1}^n to a number
that is neither 0 nor 1.