Corrected how unary negation is output
Added an extra level of parentheses around unary negation because
negation has lower precedence than exponentiation.  For example, what
was previously output as "-(a) ^ 2" is now output as "(-(a)) ^2".  As
a special case, negating a positive integer constant does not
introduce additional parentheses.  Hence, what was output as "-(3)" is
now output as "-3".